refactor(nixlbench): move RAII types into nixl_mem_region.{h,cpp}
Extract the xferFileState and NixlMemRegion RAII helpers (along with the
shared CHECK_NIXL_ERROR macro and iovListToNixlRegDlist helper they
depend on) out of nixl_worker.{h,cpp} into their own translation unit.
