+++ benchmark/nixlbench/src/worker/nixl/nixl_worker.h
@@ -31,89 +31,7 @@
 #include <nixl.h>
 #include "utils/utils.h"
 #include "worker/worker.h"
-
-struct xferFileState {
-    int fd = -1;
-    uint64_t file_size = 0;
-    uint64_t offset = 0;
-
-    xferFileState() = default;
-
-    xferFileState(int fd, uint64_t file_size, uint64_t offset) noexcept
-        : fd(fd),
-          file_size(file_size),
-          offset(offset) {}
-
-    ~xferFileState() {
-        closeFd();
-    }
-
-    xferFileState(xferFileState &&o) noexcept
-        : fd(std::exchange(o.fd, -1)),
-          file_size(o.file_size),
-          offset(o.offset) {}
-
-    xferFileState &
-    operator=(xferFileState &&o) noexcept {
-        if (this != &o) {
-            closeFd();
-            fd = std::exchange(o.fd, -1);
-            file_size = o.file_size;
-            offset = o.offset;
-        }
-        return *this;
-    }
-
-    xferFileState(const xferFileState &) = delete;
-    xferFileState &
-    operator=(const xferFileState &) = delete;
-
-private:
-    void
-    closeFd() noexcept {
-        if (fd >= 0) {
-            ::close(fd);
-        }
-    }
-};
-
-class NixlMemRegion {
-    nixlAgent *agent_ = nullptr;
-    nixlBackendH *backend_ = nullptr;
-    nixl_mem_t seg_type_ = DRAM_SEG;
-    std::vector<xferBenchIOV> iovs_;
-    std::function<void(xferBenchIOV &)> cleanup_;
-    nixl_opt_args_t cached_opt_args_;
-
-public:
-    NixlMemRegion() = default;
-    NixlMemRegion(nixlAgent *agent,
-                  nixlBackendH *backend,
-                  nixl_mem_t seg_type,
-                  std::vector<xferBenchIOV> iovs,
-                  std::function<void(xferBenchIOV &)> cleanup = nullptr);
-    ~NixlMemRegion();
-    NixlMemRegion(NixlMemRegion &&o) noexcept;
-    NixlMemRegion &
-    operator=(NixlMemRegion &&o) noexcept;
-    NixlMemRegion(const NixlMemRegion &) = delete;
-    NixlMemRegion &
-    operator=(const NixlMemRegion &) = delete;
-
-    const std::vector<xferBenchIOV> &
-    iovs() const {
-        return iovs_;
-    }
-
-    std::vector<xferBenchIOV> &
-    iovs() {
-        return iovs_;
-    }
-
-private:
-    void
-    release();
-};
+#include "worker/nixl/nixl_mem_region.h"
 
 // Use shared GusliDeviceConfig and parseGusliDeviceList declared in utils.h
